requirements.txt
langchain[all]
openai
python-dotenv
pypdf
tiktoken
ctransformers
streamlit
chromadb
pysqlite-binary
pydantic
PyPDF2
PyCryptodome
accelerate
unstructured
rapidocr_pdf[onnxruntime]
pdfminer.six
pdf2image
fitz
frontend
camelot-py
opencv-python